Edit project details sent as props in Router.jsx and refactor the Project component to change the rendered content based on the prop 'isInverted'
This commit adds a new prop, isInverted, to the Project component. If the value is true, the Project component will render a project's details on the right and the project's thumbnail on the left. If the value is false, the component will instead render a project's thumbnail on the right and the project's details on the left.