Add responsiveness using media queries
This commit updates the Project component to render a project article with a class name 'inverted' based on whether the prop 'isInverted' is true or false. This class is used to revert the positions of an inverted project's details and thumbnail so that the details are displayed before the thumbnail on smaller screens. Having all project views exactly the same improves the user's experience on smaller devices as it makes the content easier to read.